requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
blessed==1.20.0
bpython==0.24
certifi==2023.7.22
charset-normalizer==3.3.0
crispy-bootstrap5==0.7
curtsies==0.4.2
cwcwidth==0.1.9
decorator==5.1.1
Django
django-crispy-forms==2.0
django-allauth==0.57.0
executing==2.0.0
greenlet==3.0.0
idna==3.4
ipython==8.16.1
jedi==0.19.1
loguru==0.7.2
matplotlib-inline==0.1.6
numpy==1.26.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
pyxdg==0.28
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
whitenoise==6.5.0
XlsxWriter==3.1.8
wfastcgi==3.0.0
mysqlclient
django-debug-toolbar==4.2.0
faker==24.3.0
django-countries==7.6
Pillow==10.2.0